Use ShadowedRectangle for Card backgrounds
Summary:
This switches cards to use ShadowedRectangle for their background. It also refreshes their
design a bit, based on feedback from the VDG.

Still WIP because of potential design issues and I'm not entirely happy with the code for
the actual cards. Ideally I'd also move a bunch of this stuff to C++ but that'd involve a
reimplementing a bunch of Control behaviour.

BUG: 415526

Test Plan:
New cards, normal colour scheme:

{F8221005}

Breeze Dark:

{F8221010}

Reviewers: #kirigami, #vdg, cblack, mart

Reviewed By: #kirigami, #vdg, cblack, mart

Subscribers: mart, ngraham, nicolasfella, cblack, plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D28625