Introduce PlaceholderMessage component
Summary:
This patch creates the `PlaceholderMessage`, a re-usable component that can be used
to create all kinds of placeholder messages with a standardized appearance. It can be a
loading screen, an "there's no content in this view" message, an error message, and more.

Test Plan:
In Discover, with D29059: {F8251669}
In Plasma-Thunderbolt, with D29060:  {F8251670}

More will be ported if and when this patch lands.

Reviewers: #vdg, #kirigami, mart

Reviewed By: #vdg, #kirigami, mart

Subscribers: leinir, abetts, broulik, cblack, plasma-devel

Tags: #kirigami

Maniphest Tasks: T11227

Differential Revision: https://phabricator.kde.org/D29057