Add support for layers to PagePoolAction.
Summary: Currently PagePoolAction only supports pushing to a pageStack, but often it makes sense to use layers to push modal/contextual pages (outside of the primary pageStack flow). This change introduces support for a "useLayers" property that can be set on PagePoolAction, allowing it use the pageStack.layers StackView (if it exists). It also includes a unit test and a couple fixes uncovered by the test.

Test Plan: See the tst_layers unit test included.

Reviewers: #kirigami, mart

Reviewed By: #kirigami, mart

Subscribers: cblack, mart, plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D29294