Fix excessive right padding in scrollable page
Summary:
In this example
```
import QtQuick 2.5
import QtQuick.Layouts 1.1
import org.kde.kirigami 2.4 as Kirigami

Kirigami.ScrollablePage {

    Rectangle {
        width: parent.width
        height: 1000
        color: "red"
    }
}
```

the gap between the rectangle and the scrollbar was larger than the distance between the left edge of the rectangle and the window. This was caused by the extra rightPadding of the RefreshableScrollView. It also seems unnecessary to me since the ScrollView itself takes care of dodging the scroll bar (ScrollView.qml:21).

Test Plan:
Before:
{F8273998}

After:
{F8273999}

Reviewers: #kirigami, mart, ngraham

Reviewed By: #kirigami, ngraham

Subscribers: ngraham, plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D29308