Adjust ActionToolbarLayoutDetails calculations to make better use of screen real estate
Before this patch, the moreButton was always factored into calculations,
which resulted in actions hiding when there was enough screen space for them.

This patch is a bit more thoughtful on factoring moreButton into
calculations, which results in screen real estate being better used.

Items in the details now keep track of two visible states: visible with
moreButton and visible without moreButton. The states are used when
moreButton is and is not visible, with one special case: when there is
only one hidden action and the toolbar is growing. In this special case,
the moreButton is ignored in calculations.