Fix OverlaySheet closing when clicking inside layout
The `OverlaySheet` would close when clicking outside `scrollview.contentIem`.
This area doesn't include the scrollbar area, nor header or footer items, so clicking on those items,
if they do not catch the click event first, would result in an undesired closure of the sheet.

Setting the condition to check for `contentLayout` instead respect the limits of the visible card.
Clicking outside of this area still closes the sheet as expected.

BUG: 421848
FIXED-IN: 5.72