Fix import for static builds
This switches the logic for registering the Kirigami import types away from registering them explicitly, to instead register the plugin as something which can be imported. This fixes an issue in which, if you run a statically linked Kirigami 2 based application on a system with a system-installed Kirigami 2, would first look in the major-versioned location of that system installation and not in the statically linked one. In short, this avoids such clashes and hopefully results in simpler and safer use of Kirigami for static builds.