Reintroduce layer support to Page.isCurrentPage
Previously, the isCurrentPage property on Page would not return true,
unless it was the current item on the currently active layer, but
during some refactoring, this ability got lost and it would return
true in more cases (especially the layers functionality of PageRow
was used). This patch fixes the issue, by reintroducing the explicit
support for handling layers.