Force a low Z order of ShadowedRectangle's software fallback
PaintedRectangle, the item used as software rendering fallback, is an
item that gets added to the normal list of children. This means it is
part of normal Z order calculations, instead of being internal to the
item. So we need to force it into a low Z order to ensure it always acts
as "background"