Introduce PageRouter lifecycle APIs
This introduces APIs to managing the preloading and caching of pages with the PageRouter.
Notably, there are LRU caches governing the lifetime of cached and preloaded pages.
Caching and preloading are extremely popular techniques for improving the user experience of an application, as it can make apps feel significantly snappier, especially when routes take a few seconds to construct.