[passivenotification] Set explicit padding
It currently relies on Control having an implicit padding in qqc2-desktop-style (https://invent.kde.org/frameworks/qqc2-desktop-style/-/blob/master/org.kde.desktop/Control.qml#L23). PC3 Control does not have this implicit padding (and I'm not sure what the implications of adding it would be). This results in the notification not having proper padding when using the Plasma QQC2 style.

The resulting padding is 6 on my machine, which is neither smallSpacing (4) nor largeSpacing (8). I think it's better to use the existing units, but it results in a visual change. Therefore CC @teams/vdg

BUG: 419391