Account for More button size if we already know it will be visible
When there are actions with the AlwaysHide display hint set, we did not
properly account for the more button appearing, causing things to break
out of the layout. Since we know there are actions like that before we
do the layouting, we can account for that and make sure we reserve space
for the more button.