Add a placeholder image property to icon (in the style of fallback)
The name of an icon from the icon theme to show while the icon set in `source` is
being loaded. This is primarily relevant for remote sources, or those using slow-
loading image providers. The default temporary icon is `"image-x-icon"`

This will only be loaded if the source is a type which can be so long-loading
that a temporary image makes sense (e.g. a remote image, or from an ImageProvider
of the type QQmlImageProviderBase::ImageResponse)

The intention here is to ensure we aren't hardcoding things which would
also arguably be useful for the user (specifically, in cases where the
user of the component knows more precisely which type of icon is going
to be shown here, and consequently have a better choice for the
temporary image than a generic icon image).