[globaldrawer]: Explicitly set height of header
In some cases (while I haven't figured out a minimal example, this can be reproduced in Discover), the header's height calculations can just completely nope out and go to 0, causing layout breakages. Since it's not in a layout, we're free to set the heightto the sum of its content implicit heights.