Fix menus in contextualActions
If the menu contents change while the component is being created,
refresh the actions property as it's populated so that we don't miss
updates.
Otherwise they would get out of sync. This is especially a problem when
using JS arrays because QML adds them one by one, and when creating the
first we'd construct the menu and lose the subsequent ones.