Refine list separator appearance
The lines are now a little bit lighter in weight and no longer span the
full width of the list item; now they have margins equal to `largeSpacing`,
which is used for the items inside the list item itself. This makes the
lones no longer so visually objectionable, which makes them suitable for
keeping as the default appearance in list views.

Also, the top separator is removed because it is unneeded; only having one
separator at the bottom of each list item is visually sufficient, and fixes a
bug preventing you from removing the separators inside a list item when it
has a colored rectangle, since the bottom-most line is actually drawn by the
item below it.

BUG: 428739
FIXED-IN: 5.77